refactor: split relay roles
