+++ src/bluetooth_2_usb/extended_mouse.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 from .logging import get_logger
 
-_logger = get_logger()
+logger = get_logger(__name__)
 
 
 def _clamp_hid_i8(value: int) -> int:
@@ -66,7 +66,7 @@ class ExtendedMouse:
             self.report[3:5] = partial_y.to_bytes(2, "little", signed=True)
             self.report[5] = partial_wheel & 0xFF
             self.report[6] = partial_pan & 0xFF
-            _logger.debug(
+            logger.debug(
                 "Sending mouse movement to gadget: buttons=0x%02x x=%s y=%s "
                 "wheel=%s pan=%s report=%s",
                 self.report[0],
